refactor(loaders): move to a common folder
Update demos and templates.
