+++ loaders/css2json-loader.js
