+++ loaders/markup-hot-loader.js
