+++ loaders/moduleid-compat-loader.js
