+++ loaders/script-hot-loader.js
