+++ plugins/NativeScriptSnapshotPlugin/index.js
@@ -48,8 +48,8 @@ exports.NativeScriptSnapshotPlugin = (function() {
             snapshotEntryContent += `
             require("${
                 options.angular ?
-                    'nativescript-dev-webpack/load-application-css-angular' :
-                    'nativescript-dev-webpack/load-application-css-regular'
+                    'nativescript-dev-webpack/loaders/load-application-css-angular' :
+                    'nativescript-dev-webpack/loaders/load-application-css-regular'
             }")();
             `;
         }
