+++ templates/webpack.angular.js
@@ -152,12 +152,12 @@ module.exports = env => {
                     use: [
                         // Require all Android app components
                         platform === "android" && {
-                            loader: "nativescript-dev-webpack/android-app-components-loader",
+                            loader: "nativescript-dev-webpack/loaders/android-app-components-loader",
                             options: { modules: appComponents }
                         },
 
                         {
-                            loader: "nativescript-dev-webpack/bundle-config-loader",
+                            loader: "nativescript-dev-webpack/loaders/bundle-config-loader",
                             options: {
                                 angular: true,
                                 loadCss: !snapshot, // load the application css if in debug mode
@@ -191,7 +191,7 @@ module.exports = env => {
                 {
                     test: /(?:\.ngfactory\.js|\.ngstyle\.js|\.ts)$/,
                     use: [
-                        "nativescript-dev-webpack/moduleid-compat-loader",
+                        "nativescript-dev-webpack/loaders/moduleid-compat-loader",
                         "@ngtools/webpack",
                     ]
                 },
